+++ src/pages/MainPage.jsx
@@ -1,6 +1,6 @@
-import CouponBanner from 'components/CouponBanner';
-import ReviewSlide from 'components/ReviewSlide';
-import SubscribeBanner from 'components/SubscribeBanner';
+import CouponBanner from 'components/main/CouponBanner';
+import ReviewSlide from 'components/main/ReviewSlide';
+import SubscribeBanner from 'components/main/SubscribeBanner';
 import DetailIntro from 'components/main/DetailIntro';
 import IntroLayout from 'components/main/IntroLayout';
 import MainBanner from 'components/main/MainBanner';
